// src/app/model/location.ts
export class Location {
  ID: number;
  ClientID: number;
  LocationCode: string;
  LocationName: string;
  LocationSPOC_Name: string;
  BillingAddress: string;
  ShippingAddress: string;
  GST_Number: string;
  Status: number;
  LastUpdatedOn: Date;
  LastUpdatedBy: number;
  SPOC_Email: string;
  SPOC_ContactNumber: string;
  ClientName: string;
  InvoiceCentreId: number;
  InvoiceCentre: string;
}