// src/app/model/BusinessPartner.ts
export class BusinessPartner {
  Id: number;
  BusinessPartnerCode: string;
  BusinessPartnerName: string;
  BusinessPartnerSPOC: string;
  BPReferenceID: number;
  Location: string;
  Address: number;
  GSTNumber: string;
  CommencementDate: Date;
  RenewalDate: Date;
  Status: string;
  LastUpdatedOn: Date;
  LastUpdatedBy: number;
}