// src/app/model/CityBusinessPartner.ts
export class CityBusinessPartner {
  Id: number;
  CityBusinessPartnerCode: string;
  CityBusinessPartnerName: string;
  CityBusinessPartnerSPOC: string;
  CBPReferenceID: number;
  Location: string;
  Address: number;
  GSTNumber: string;
  CommencementDate: Date;
  RenewalDate: Date;
  Status: string;
  LastUpdatedOn: Date;
  LastUpdatedBy: number;
}